Add version 2.1c
Changelog:

2.1c - 2016-12-22
Geographic Region support, RPI3, new install image.
Also correct webserver behaviour for insecure frames on
https pages by downgrading when required. This does not
result in configuration details being in plain text as the
configuration page should upgrade to https for the process.

2.1 FE - 2016-12-20
Front end changes to handle new configuration fields for v0.2.6,
region selection based on location of pin

2.1b - (not released)
Correct handling of IP and MAC addresses on devices where locale is not English.

2.1a - 2016-11-02
Merging in changes for latest RPis (Jessie thanks to David Uursem).
NOTE: This is just to test Jessie and should not be used live yet!
- Jessie support
- Detect pae unames and use this to install 32bit software
- Use MD5 SUMs from OGN Download server, this allows automated upgrade and downgrade if pushed by OGN team.
- 2.0 installations will not automatically upgrade to this version yet.

This version was fetched from ognconfig.onglide.com on 2017-01-10.